requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
pymysql==1.1.0
cryptography==41.0.7

